Update a Nível de Segurança e Detalhes
Foram realizadas alterações que impedem que utilizadores que não o admin alterem/editem o conteúdo do menu dinãmico "Sobre Techn Art".Também já não é possível copiar um link de um browser pro outro, evitando assim o chamado "hotlinking".Por último foi revisto o perfil de utilizador "Todos" ,uma vez que, antigamente quando o admin  associava um documento ao perfil "Todos" nenhum tipo de user seja colaborador, integrado, etc, conseguia ter acesso ao documento através da sua conta, tal já não acontece agora.

+++ backoffice/missao/index.php
@@ -71,7 +71,11 @@ $totalPages = ceil($totalRows / $perPage);
 <div class="container mt-4">
     <div class="table-title d-flex justify-content-between align-items-center">
         <h2>Missão e Objetivos</h2>
-        <a href="add.php" class="btn btn-success"><i class="fas fa-plus"></i> Adicionar Novo</a>
+       <?php if ($_SESSION["autenticado"] == 'administrador'): ?>
+            <a href="add.php" class="btn btn-success"><i class="fas fa-plus"></i> Adicionar Novo</a>
+       <?php else: ?>
+            <span class="text-muted">Acesso Restrito</span>
+        <?php endif; ?>
     </div>
 
 
@@ -96,9 +100,13 @@ $totalPages = ceil($totalRows / $perPage);
                         <td><textarea class="form-control" readonly><?= $row['titulo_pt'] ?></textarea></td>
                         <td><textarea class="form-control" readonly><?= $row['titulo_en'] ?></textarea></td>
                         <td style="min-width:200px;">
-                            <a href="edit.php?id=<?= $row['id'] ?>" class="w-100 mb-1 btn btn-primary">Alterar</a>
-                            <a href="delete.php?id=<?= $row['id'] ?>" class="w-100 mb-1 btn btn-danger">Apagar</a>
-                        </td>
+                            <!-- Verificação de permissões e seguranaça -->
+                            <?php if ($_SESSION["autenticado"] == 'administrador' || $_SESSION["autenticado"] == $row["id"]): ?>
+                                <a href="edit.php?id=<?= $row['id'] ?>" class="w-100 mb-1 btn btn-primary">Alterar</a>
+                                <a href="delete.php?id=<?= $row['id'] ?>" class="w-100 mb-1 btn btn-danger">Apagar</a>
+                            <?php else: ?>
+                                <span class="text-muted w-100 mb-1">Acesso Restrito</span>
+                            <?php endif; ?>
                     </tr>
                 <?php endwhile; ?>
             </tbody>
